requirements.txt
pandas
numpy
yfinance
pandas-datareader
tensorflow
scikit-learn
joblib
apache-airflow
sqlalchemy
psycopg2-binary
python-dotenv